+++ src/tools/plugins.ts
@@ -11,3 +11,4 @@ export const jwtCreate = new Elysia()
 
 export const cookieExpose = new Elysia()
   .use(cookie())
+
